moving PgTitle into a template sub-folder under components
