+++ src/components/template/PgTitle/types.ts
